Ajout authentification sur les routes de courses

+++ routes/course.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const request = require('../database/course.js');
+const requestUser = require('../database/utilisateurs.js');
 const router = express.Router();
 const auth = require('../middleware/authentification')
 
@@ -7,6 +8,10 @@ router.get('/:id_course', async (req, res) => {
     try {
         const id_course = req.params.id_course
         const data = await request.getCourse(id_course)
+
+        // Verifie si la course appartient a l'utilisateur
+        if (req.user.id != data[0].id_utilisateur) return res.status(401).json({ message: "Unauthorized."})
+
 
         if (data.length === 0) {
             return res.status(404).json({ message: 'Aucune course avec cet id trouvée' });
@@ -20,19 +25,21 @@ router.get('/:id_course', async (req, res) => {
 
 router.post('/', auth, async (req, res) => {
     try {
-        const jwt_user = req.user
         const id_utilisateur = req.body.id_utilisateur
         const distance = req.body.distance || null
-        //const duree = req.body.duree || null
-        const duree = new Date(0, 0, 0, 1, 10, 5, 0)
-        const date = Date.now().toString()
+        const duree = req.body.duree || null
+        const date = new Date().toISOString()
 
-        console.log(date, distance, duree, id_utilisateur)
-        if (jwt_user.id != id_utilisateur) return res.status(401).json({ message: "Unauthorized."})
+        // Authentification de l'Utilisateur
+        if (req.user.id != id_utilisateur) return res.status(401).json({ message: "Unauthorized."})
+        
+        // Verifie si l'Utilisateur est bien dans la base de donnees
+        const user = await requestUser.getUserById(id_utilisateur)
+        if (!user) {
+            return res.status(404).json({ message: "Aucun utilisateur avec cet ID trouvé." })    
+        }
 
         const data = await request.addCourse(date, distance, duree, id_utilisateur)
-        
-        console.log(data)
         
         res.status(200).json({ message: "success" })
     } catch (error) {
@@ -40,12 +47,17 @@ router.post('/', auth, async (req, res) => {
     }
 })
 
-router.delete('/:id_course', async (req, res) => {
+router.delete('/:id_course', auth, async (req, res) => {
     try {
         const id_course = req.params.id_course
-        const data = await request.deleteCourse(id_course)
         
-        console.log(data)
+        // Verifie si la course appartient a l'utilisateur
+        const course = await request.getCourse(id_course)
+        if (req.user.id != course[0].id_utilisateur) return res.status(401).json({ message: "Unauthorized."})
+
+        //Suppression de la course
+        const data = await request.deleteCourse(id_course)
+
         if (data.length === 0) {
             return res.status(404).json({ message: 'Aucune course avec cet id trouvée' });
         }
